Original change: http://codereview.chromium.org/4856003/ (landing on behalf of kristianm)
Suppress Android warnings

Android generates warnings for comparing signed and unsigned numbers.
The ones in this CL should be safe.

BUG=62949
TEST=None

Review URL: http://codereview.chromium.org/5035002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@66420 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/base/net_util.cc
@@ -2030,12 +2030,12 @@ bool IPNumberMatchesPrefix(const IPAddressNumber& ip_number,
 uint16* GetPortFieldFromAddrinfo(const struct addrinfo* info) {
   DCHECK(info);
   if (info->ai_family == AF_INET) {
-    DCHECK_EQ(sizeof(sockaddr_in), info->ai_addrlen);
+    DCHECK_EQ(sizeof(sockaddr_in), static_cast<size_t>(info->ai_addrlen));
     struct sockaddr_in* sockaddr =
         reinterpret_cast<struct sockaddr_in*>(info->ai_addr);
     return &sockaddr->sin_port;
   } else if (info->ai_family == AF_INET6) {
-    DCHECK_EQ(sizeof(sockaddr_in6), info->ai_addrlen);
+    DCHECK_EQ(sizeof(sockaddr_in6), static_cast<size_t>(info->ai_addrlen));
     struct sockaddr_in6* sockaddr =
         reinterpret_cast<struct sockaddr_in6*>(info->ai_addr);
     return &sockaddr->sin6_port;
